+++ src/lib.rs
@@ -1,8 +1,5 @@
-// WASM cannot be executed through the test runner and we need to avoid building
-// WASM code for non-WASM targets. We can solve both of these with a cfg flag.
-
-#[cfg(not(test))]
+#[cfg(feature = "wasm")]
 mod proto;
 
-#[cfg(not(test))]
+#[cfg(feature = "wasm")]
 pub use proto::*;